Fleshed out model training with Lightning ⚡
