+++ src/models/palette/network.py
@@ -1,21 +1,59 @@
 import math
 import torch
+import torch.nn as nn
 from inspect import isfunction
 from functools import partial
 import numpy as np
 from tqdm import tqdm
-from core.base_network import BaseNetwork
 
 
-class Network(BaseNetwork):
+class Network(nn.Module):
     def __init__(self, unet, beta_schedule, init_type='kaiming', gain=0.02):
-        super(Network, self).__init__(init_type, gain)
-        
+        super(Network, self).__init__()
         self.denoise_fn = unet
         self.beta_schedule = beta_schedule
+        self.init_type = init_type
+        self.gain = gain
 
-    def set_loss(self, loss_fn):
-        self.loss_fn = loss_fn
+    def init_weights(self):
+        """
+        initialize network's weights
+        init_type: normal | xavier | kaiming | orthogonal
+        https://github.com/junyanz/pytorch-CycleGAN-and-pix2pix/blob/9451e70673400885567d08a9e97ade2524c700d0/models/networks.py#L39
+        """
+
+        def init_func(m):
+            classname = m.__class__.__name__
+            if classname.find('InstanceNorm2d') != -1:
+                if hasattr(m, 'weight') and m.weight is not None:
+                    nn.init.constant_(m.weight.data, 1.0)
+                if hasattr(m, 'bias') and m.bias is not None:
+                    nn.init.constant_(m.bias.data, 0.0)
+            elif hasattr(m, 'weight') and (classname.find('Conv') != -1 or classname.find('Linear') != -1):
+                if self.init_type == 'normal':
+                    nn.init.normal_(m.weight.data, 0.0, self.gain)
+                elif self.init_type == 'xavier':
+                    nn.init.xavier_normal_(m.weight.data, gain=self.gain)
+                elif self.init_type == 'xavier_uniform':
+                    nn.init.xavier_uniform_(m.weight.data, gain=1.0)
+                elif self.init_type == 'kaiming':
+                    nn.init.kaiming_normal_(m.weight.data, a=0, mode='fan_in')
+                elif self.init_type == 'orthogonal':
+                    nn.init.orthogonal_(m.weight.data, gain=self.gain)
+                elif self.init_type == 'none':  # uses pytorch's default init method
+                    m.reset_parameters()
+                else:
+                    raise NotImplementedError(
+                        f'initialization method {self.init_type} is not implemented'
+                    )
+                if hasattr(m, 'bias') and m.bias is not None:
+                    nn.init.constant_(m.bias.data, 0.0)
+
+        self.apply(init_func)
+        # propagate to children
+        for m in self.children():
+            if hasattr(m, 'init_weights'):
+                m.init_weights(self.init_type, self.gain)
 
     def set_new_noise_schedule(self, device):
 
@@ -23,8 +61,7 @@ class Network(BaseNetwork):
 
         # betas = make_beta_schedule(**self.beta_schedule[phase])
         betas = make_beta_schedule(**self.beta_schedule)
-        betas = betas.detach().cpu().numpy() if isinstance(
-            betas, torch.Tensor) else betas
+        betas = betas.detach().cpu().numpy() if isinstance(betas, torch.Tensor) else betas
         alphas = 1. - betas
 
         timesteps, = betas.shape
@@ -34,21 +71,41 @@ class Network(BaseNetwork):
         gammas_prev = np.append(1., gammas[:-1])
 
         # calculations for diffusion q(x_t | x_{t-1}) and others
-        self.register_buffer('gammas', to_torch(gammas))
-        self.register_buffer('sqrt_recip_gammas',
-                             to_torch(np.sqrt(1. / gammas)))
-        self.register_buffer('sqrt_recipm1_gammas',
-                             to_torch(np.sqrt(1. / gammas - 1)))
+        self.register_buffer(
+            'gammas',
+            to_torch(gammas)
+        )
+        self.register_buffer(
+            'sqrt_recip_gammas',
+            to_torch(np.sqrt(1. / gammas))
+        )
+        self.register_buffer(
+            'sqrt_recipm1_gammas',
+            to_torch(np.sqrt(1. / gammas - 1))
+        )
 
         # calculations for posterior q(x_{t-1} | x_t, x_0)
         posterior_variance = betas * (1. - gammas_prev) / (1. - gammas)
+
         # below: log calculation clipped because the posterior variance is 0 at the beginning of the diffusion chain
-        self.register_buffer('posterior_log_variance_clipped', to_torch(
-            np.log(np.maximum(posterior_variance, 1e-20))))
-        self.register_buffer('posterior_mean_coef1', to_torch(
-            betas * np.sqrt(gammas_prev) / (1. - gammas)))
-        self.register_buffer('posterior_mean_coef2', to_torch(
-            (1. - gammas_prev) * np.sqrt(alphas) / (1. - gammas)))
+        self.register_buffer(
+            'posterior_log_variance_clipped',
+            to_torch(
+                np.log(np.maximum(posterior_variance, 1e-20))
+            )
+        )
+        self.register_buffer(
+            'posterior_mean_coef1',
+            to_torch(
+                betas * np.sqrt(gammas_prev) / (1. - gammas)
+            )
+        )
+        self.register_buffer(
+            'posterior_mean_coef2',
+            to_torch(
+                (1. - gammas_prev) * np.sqrt(alphas) / (1. - gammas)
+            )
+        )
 
     def predict_start_from_noise(self, y_t, t, noise):
         return (
@@ -61,20 +118,30 @@ class Network(BaseNetwork):
             extract(self.posterior_mean_coef1, t, y_t.shape) * y_0_hat +
             extract(self.posterior_mean_coef2, t, y_t.shape) * y_t
         )
+
         posterior_log_variance_clipped = extract(
             self.posterior_log_variance_clipped, t, y_t.shape)
+
         return posterior_mean, posterior_log_variance_clipped
 
     def p_mean_variance(self, y_t, t, clip_denoised: bool, y_cond=None):
         noise_level = extract(self.gammas, t, x_shape=(1, 1)).to(y_t.device)
+
         y_0_hat = self.predict_start_from_noise(
-            y_t, t=t, noise=self.denoise_fn(torch.cat([y_cond, y_t], dim=1), noise_level))
+            y_t,
+            t=t,
+            noise=self.denoise_fn(torch.cat([y_cond, y_t], dim=1), noise_level)
+        )
 
         if clip_denoised:
             y_0_hat.clamp_(-1., 1.)
 
         model_mean, posterior_log_variance = self.q_posterior(
-            y_0_hat=y_0_hat, y_t=y_t, t=t)
+            y_0_hat=y_0_hat,
+            y_t=y_t,
+            t=t
+        )
+
         return model_mean, posterior_log_variance
 
     def q_sample(self, y_0, sample_gammas, noise=None):
@@ -87,33 +154,54 @@ class Network(BaseNetwork):
     @torch.no_grad()
     def p_sample(self, y_t, t, clip_denoised=True, y_cond=None):
         model_mean, model_log_variance = self.p_mean_variance(
-            y_t=y_t, t=t, clip_denoised=clip_denoised, y_cond=y_cond)
+            y_t=y_t,
+            t=t,
+            clip_denoised=clip_denoised,
+            y_cond=y_cond
+        )
+
         noise = torch.randn_like(y_t) if any(t > 0) else torch.zeros_like(y_t)
+
         return model_mean + noise * (0.5 * model_log_variance).exp()
 
     @torch.no_grad()
     def restoration(self, y_cond, y_t=None, y_0=None, mask=None, sample_num=8):
         b, *_ = y_cond.shape
 
-        assert self.num_timesteps > sample_num, 'num_timesteps must greater than sample_num'
+        assert self.num_timesteps > sample_num, 'num_timesteps must be greater than sample_num'
         sample_inter = (self.num_timesteps//sample_num)
 
         y_t = default(y_t, lambda: torch.randn_like(y_cond))
         ret_arr = y_t
-        for i in tqdm(reversed(range(0, self.num_timesteps)), desc='sampling loop time step', total=self.num_timesteps):
+
+        for i in tqdm(
+            reversed(range(0, self.num_timesteps)),
+            desc='sampling loop time step',
+            total=self.num_timesteps
+        ):
+
             t = torch.full((b,), i, device=y_cond.device, dtype=torch.long)
+
             y_t = self.p_sample(y_t, t, y_cond=y_cond)
+
             if mask is not None:
                 y_t = y_0*(1.-mask) + mask*y_t
+
             if i % sample_inter == 0:
                 ret_arr = torch.cat([ret_arr, y_t], dim=0)
+
         return y_t, ret_arr
 
     def forward(self, y_0, y_cond=None, mask=None, noise=None):
         # sampling from p(gammas)
         b, *_ = y_0.shape
-        t = torch.randint(1, self.num_timesteps, (b,),
-                          device=y_0.device).long()
+        t = torch.randint(
+            1,
+            self.num_timesteps,
+            (b,),
+            device=y_0.device
+        ).long()
+
         gamma_t1 = extract(self.gammas, t-1, x_shape=(1, 1))
         sqrt_gamma_t2 = extract(self.gammas, t, x_shape=(1, 1))
         sample_gammas = (sqrt_gamma_t2-gamma_t1) * \
@@ -121,27 +209,36 @@ class Network(BaseNetwork):
         sample_gammas = sample_gammas.view(b, -1)
 
         noise = default(noise, lambda: torch.randn_like(y_0))
+
         y_noisy = self.q_sample(
-            y_0=y_0, sample_gammas=sample_gammas.view(-1, 1, 1, 1), noise=noise)
+            y_0=y_0,
+            sample_gammas=sample_gammas.view(-1, 1, 1, 1),
+            noise=noise
+        )
 
         if mask is not None:
             noise_hat = self.denoise_fn(
-                torch.cat([y_cond, y_noisy*mask+(1.-mask)*y_0], dim=1), sample_gammas)
+                torch.cat(
+                    [y_cond, y_noisy*mask+(1.-mask)*y_0],
+                    dim=1
+                ),
+                sample_gammas
+            )
             loss = self.loss_fn(mask*noise, mask*noise_hat)
         else:
             noise_hat = self.denoise_fn(
-                torch.cat([y_cond, y_noisy], dim=1), sample_gammas)
+                torch.cat(
+                    [y_cond, y_noisy],
+                    dim=1
+                ),
+                sample_gammas
+            )
             loss = self.loss_fn(noise, noise_hat)
         return loss
 
 
-# gaussian diffusion trainer class
-def exists(x):
-    return x is not None
-
-
 def default(val, d):
-    if exists(val):
+    if val is not None:
         return val
     return d() if isfunction(d) else d
 
@@ -150,8 +247,6 @@ def extract(a, t, x_shape=(1, 1, 1, 1)):
     b, *_ = t.shape
     out = a.gather(-1, t)
     return out.reshape(b, *((1,) * (len(x_shape) - 1)))
-
-# beta_schedule function
 
 
 def _warmup_beta(linear_start, linear_end, n_timestep, warmup_frac):
@@ -164,22 +259,42 @@ def _warmup_beta(linear_start, linear_end, n_timestep, warmup_frac):
 
 def make_beta_schedule(schedule, n_timestep, linear_start=1e-6, linear_end=1e-2, cosine_s=8e-3):
     if schedule == 'quad':
-        betas = np.linspace(linear_start ** 0.5, linear_end ** 0.5,
-                            n_timestep, dtype=np.float64) ** 2
+        betas = np.linspace(
+            linear_start ** 0.5,
+            linear_end ** 0.5,
+            n_timestep,
+            dtype=np.float64
+        ) ** 2
     elif schedule == 'linear':
-        betas = np.linspace(linear_start, linear_end,
-                            n_timestep, dtype=np.float64)
+        betas = np.linspace(
+            linear_start,
+            linear_end,
+            n_timestep,
+            dtype=np.float64
+        )
     elif schedule == 'warmup10':
-        betas = _warmup_beta(linear_start, linear_end,
-                             n_timestep, 0.1)
+        betas = _warmup_beta(
+            linear_start,
+            linear_end,
+            n_timestep,
+            0.1
+        )
     elif schedule == 'warmup50':
-        betas = _warmup_beta(linear_start, linear_end,
-                             n_timestep, 0.5)
+        betas = _warmup_beta(
+            linear_start,
+            linear_end,
+            n_timestep,
+            0.5
+        )
     elif schedule == 'const':
         betas = linear_end * np.ones(n_timestep, dtype=np.float64)
     elif schedule == 'jsd':  # 1/T, 1/(T-1), 1/(T-2), ..., 1
-        betas = 1. / np.linspace(n_timestep,
-                                 1, n_timestep, dtype=np.float64)
+        betas = 1. / np.linspace(
+            n_timestep,
+            1,
+            n_timestep,
+            dtype=np.float64
+        )
     elif schedule == "cosine":
         timesteps = (
             torch.arange(n_timestep + 1, dtype=torch.float64) /
