+++ src/models/train_model.py
@@ -1,19 +1,16 @@
-import yaml
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import WandbLogger
+from pytorch_lightning import Trainer
 from pytorch_lightning import seed_everything
-import pytorch_lightning as pl
-from palette.model import Palette
-import utils
-import torch
+import lightning_model as lm
+import lightning_data as ld
+import yaml
+
 
 # Configs
 config = yaml.safe_load(open('./config.yaml'))
 model_config = config['training']['model_config_path']
 model_config = yaml.safe_load(open(model_config))
-
-#   Model args
-paths: dict[str, str]           = config['models']
 
 #   Unet
 unet_name: str                  = model_config['unet']['name']
@@ -24,87 +21,84 @@ ema_scheduler: dict             = model_config['ema_scheduler']['ema_scheduler_a
 if not model_config['ema_scheduler']['use']:
     ema_scheduler = None
 
+#   Optimizer
 optimizer_name: str             = model_config['optimizer']['name']
 optimizer_args: dict            = model_config['optimizer']['optimizer_args']
+
+#   Scheduler
 scheduler_args: dict            = model_config['scheduler_args']
 
-# Misc
-seed: int                       = config['misc']['seed']
+#   Loss
+loss_name: str                  = model_config['loss']['loss_name']
 
-# Training
+#   Training
 training_args: dict             = config['training']['training_args']
+
+#   Trainer config
+trainer_args: dict              = config['training']['trainer']
 
 #   Dataloader
 batch_size: int                 = config['training']['dataloader']['batch_size']
 dataloader_workers: int         = config['training']['dataloader']['dataloader_workers']
 
-# Dataset
+#   Dataset
 images_path: str                = config['dataset']['images_dir']
 annotations_path: str           = config['dataset']['annotations_path']
 
-#   Trainer config
-# trainer_args: dict              = config['training']['trainer']
+#   Callbacks
+ckpt_callback_args: dict        = config['callbacks']['model_checkpoint']
+
+#   Misc things
+paths: dict                     = config['paths']
+seed: int                       = config['misc']['seed']
+
+#   Model init
+model = lm.Diffusar(
+    unet_name=unet_name, 
+    unet_args=unet_args,
+    scheduler_args=scheduler_args, 
+    ema_scheduler=ema_scheduler,
+    optimizer_name=optimizer_name,
+    optimizer_args=optimizer_args,
+    loss_name=loss_name,
+    batch_size=batch_size
+)
+
+#   Datamodule init
+datamodule = ld.DiffusarData(
+    train_images_path=images_path,
+    train_annotations_path=annotations_path,
+    batch_size=batch_size,
+    dataloader_workers=dataloader_workers
+)
+
+#   Wandb logger callback
+logger = WandbLogger(
+    log_model=config['wandb']['log_model'],
+    project=config['wandb']['project_name'],
+    name=config['wandb']['name']
+)
+
+logger.watch(
+    model=model, 
+    log=config['wandb']['log'],
+    log_freq=config['wandb']['log_freq']
+)
 
 #   Callbacks
-# ckpt_callback_args: dict        = config['callbacks']['model_checkpoint']
+checkpoint_callback = ModelCheckpoint(
+    **ckpt_callback_args
+)
 
-# Logger callback
-# logger = WandbLogger(
-#     log_model=config['wandb']['log_model'],
-#     project=config['wandb']['project_name'],
-#     name=config['wandb']['name']
-# )
-
-# logger.watch(
-#     model, 
-#     log=config['wandb']['log'],
-#     log_freq=config['wandb']['log_freq']
-# )
-
-# Callbacks
-# checkpoint_callback = ModelCheckpoint(
-#     **ckpt_callback_args
-# )
-
-# Trainer
-# trainer = pl.Trainer(
-#     logger=logger,
-#     callbacks=checkpoint_callback,
-#     **trainer_args
-# )
+#   Trainer
+trainer = Trainer(
+    logger=logger,
+    callbacks=checkpoint_callback,
+    **trainer_args
+)
 
 # Train!
-# trainer.fit(model, dataloader)
-
-
-
-def main():
-    # torch.backends.cudnn.enabled = True
-    seed_everything(seed)
-
-    network = utils.get_network(unet_name, unet_args, scheduler_args)
-    optimizer = utils.get_optimizer(network, optimizer_name, optimizer_args)
-    loss_fn = utils.get_loss_fn()
-    dataloader = utils.get_dataloader(images_path, annotations_path, dataloader_workers, batch_size)
-
-    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
-    model = Palette(
-        network = network, 
-        loss = loss_fn,
-        optimizer=optimizer,
-        main_loader=dataloader,
-        val_loader=None,
-        ema_scheduler=ema_scheduler,
-        phase = 'train',
-        device = device,
-        **training_args,
-        **paths
-    )
-
-    model.train()
-    # print(model.netG_EMA)
-
-
 if __name__ == '__main__':
-    main()
+    seed_everything(seed)
+    model.train()
+    trainer.fit(model=model, datamodule=datamodule)
